Redesigned gestures module to header-only
